web_requirements.txt
Django>=3.0<4.0
PyYAML==5.3